requirements.txt
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asynch==0.2.3
asyncpg==0.29.0
certifi==2023.11.17
charset-normalizer==3.3.2
ciso8601==2.3.1
click==8.1.7
clickhouse-connect==0.5.23
clickhouse-driver==0.2.6
clickhouse-sqlalchemy==0.3.0
fastapi==0.108.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
iso8601==2.1.0
leb128==1.0.5
lz4==4.3.3
Mako==1.3.0
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv
pytest==7.4.4
python-multipart==0.0.6
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.32.0.post1
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.25.0
zstandard==0.22.0
zstd==1.5.5.1
